requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
Django==3.2.7
django-filter==2.4.0
django-opentracing==1.1.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
jaeger-client==4.8.0
jsonfield==3.1.0
mysqlclient==2.0.3
opentracing==2.4.0
phonenumbers==8.12.31
promise==2.3
pytz==2021.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.2.4
text-unidecode==1.3
threadloop==1.0.2
thrift==0.13.0
tornado==6.1

djongo~=1.3.6